requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.35.2
alpaca-py==0.13.3
pandas==2.2.1
numpy==1.26.4
sqlalchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0
websockets>=11.0.3,<12.0.0
ta==0.11.0  # Technical analysis library
pytest==8.0.2  # For testing
python-dateutil==2.8.2
pytz==2024.1 